force conversion to double in display_vec and addeb back armadillo libray in linking.

+++ include/hostcode.hpp
@@ -86,7 +86,7 @@ void linspace(double min, double max, int N, double* grid);
 template <class T>
 void display_vec(T vec, int size) {
 	for (int i = 0; i < size; i++) {
-		std::printf("The %ith element, @[%i] = %f\n", i+1, i, vec[i]);
+		std::printf("The %ith element, @[%i] = %f\n", i+1, i, (double) vec[i]);
 	};
 };
 
@@ -95,7 +95,7 @@ template <class T>
 void display_vec(T vec) {
 	int size = vec.size();
 	for (int i = 0; i < size; i++) {
-		std::printf("The %ith element, @[%i] = %f\n", i+1, i,vec[i]);
+		std::printf("The %ith element, @[%i] = %f\n", i+1, i, (double) vec[i]);
 	};
 };
